+++ packages/c2pa-web/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @contentauth/c2pa-web
 
+## 0.9.0
+
+### Minor Changes
+
+- 39ea594: Update settings resolution logic and functions to properly merge with base settings.
+
+## 0.8.4
+
+### Patch Changes
+
+- fd4351f: Bump to latest c2pa-rs version
+- Updated dependencies [fd4351f]
+  - @contentauth/c2pa-types@0.4.6
+  - @contentauth/c2pa-wasm@0.6.3
+
 ## 0.8.3
 
 ### Patch Changes